Fixed the spawing of enemies and of the player, made it be dependent on the tilemap

+++ game/healthbars/enemy_healthbar.py
@@ -14,7 +14,7 @@ class EnemyHealthBar(pygame.sprite.Sprite):
         self.rect = self.image.get_rect(topleft=(x, y))
         self.update_image()
 
-    def update_details(self, x: int, y: int, curr_hp: int):
+    def update_details(self, x: float, y: float, curr_hp: int):
         self.curr_hp = curr_hp
         self.rect.topleft = (x + HEALTHBAR_OFFSET_X, y)
